Encapsulate Firestore code in own module.
